* Added more logging * Other cleanups
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzauthd@18025 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ test/gcrypt.cpp
@@ -27,7 +27,7 @@ void test_gcrypt()
     // decompose the key into values that can be sent in packets
     size_t n_len;
     uint8 *key_n;
-    uint32 e;
+    uint16 e;
     if(!sRSAManager.getPublicKey().getValues(key_n, n_len, e)) return;
     if(e != 65537) return;
